requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
DateTime==5.5
Django==5.0.7
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
flower==2.0.1
humanize==4.10.0
idna==3.7
joblib==1.4.2
kombu==5.4.1
mysqlclient==2.2.4
numpy==2.1.0
packaging==24.1
pandas==2.2.2
plotly==5.23.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.8
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
tenacity==9.0.0
threadpoolctl==3.5.0
tornado==6.4.1
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.0.3
